Always use symbol for ivar names
This commit switches `Memoizer.ivar_name` over to returning a symbol
instead of a string, and to include the `@`. All of the methods for
accessing instance variables (get, set, defined?) convert this value to
a symbol when a string is passed in, and we're no longer using the
string value in Memoizer.