Improve support for ruby2-style keyword arguments
This commit adds support for ruby2-style keyword arguments that will be
compatible with ruby 2.6, 2.7, and 3.0 without errors or warnings.
Although this is currently being done in a way that is compatible with
ruby 2.6, users of this gem should expect that `ruby2_keywords` will be
removed after ruby 2.6 hits EOL.

To verify that it is compatible with all versions, tests have been added
for many new combinations of hash and keyword arguments.